Removing old code
