--- common/src/commonMain/kotlin/me/kevinschildhorn/common/deprecated/atomic/atoms/Atom.kt
@@ -1,7 +1,0 @@
-package me.kevinschildhorn.common.deprecated.atomic.atoms
-
-import me.kevinschildhorn.common.deprecated.atomic.atoms.interfaces.AtomType
-
-abstract class Atom {
-    abstract val type: AtomType
-}
