--- common/src/commonMain/kotlin/me/kevinschildhorn/common/deprecated/atomic/atoms/interfaces/ResourceAtom.kt
@@ -1,5 +1,0 @@
-package me.kevinschildhorn.common.deprecated.atomic.atoms.interfaces
-
-interface ResourceAtom<T> {
-    val image: T
-}
