--- common/src/commonMain/kotlin/me/kevinschildhorn/common/deprecated/atomic/molecules/BaseMolecule.kt
@@ -1,7 +1,0 @@
-package me.kevinschildhorn.common.deprecated.atomic.molecules
-
-import me.kevinschildhorn.common.deprecated.atomic.atoms.Atom
-
-abstract class BaseMolecule {
-    abstract val atoms: List<Atom>
-}
