--- common/src/commonMain/kotlin/me/kevinschildhorn/common/deprecated/businesslogic/CreatingProfileState.kt
@@ -1,9 +1,0 @@
-package me.kevinschildhorn.common.deprecated.businesslogic
-
-enum class CreatingProfileState {
-    NotReady,
-    Ready,
-    Loading,
-    Complete,
-    Error,
-}
