--- common/src/commonMain/kotlin/me/kevinschildhorn/common/deprecated/businesslogic/validation/BaseValidationState.kt
@@ -1,5 +1,0 @@
-package me.kevinschildhorn.common.deprecated.businesslogic.validation
-
-interface BaseValidationState {
-    val isValid: Boolean
-}
