--- common/src/commonMain/kotlin/me/kevinschildhorn/common/deprecated/businesslogic/validation/PasswordValidator.kt
@@ -1,16 +1,0 @@
-package me.kevinschildhorn.common.deprecated.businesslogic.validation
-
-object PasswordValidator {
-
-    private val letterRegex = ".*[a-zA-Z].*".toRegex()
-    private val numberRegex = ".*[\\d].*".toRegex()
-
-    fun verifyPassword(password: String): PasswordValidationState =
-        when {
-            password.isEmpty() -> PasswordValidationState.Empty
-            !letterRegex.matches(password) -> PasswordValidationState.NoLetter
-            !numberRegex.matches(password) -> PasswordValidationState.NoNumber
-            password.length < 10 -> PasswordValidationState.NotTenCharacters
-            else -> PasswordValidationState.Valid
-        }
-}
