--- common/src/commonMain/kotlin/me/kevinschildhorn/common/deprecated/uilogic/enums/TextFieldValidationState.kt
@@ -1,11 +1,0 @@
-package me.kevinschildhorn.common.deprecated.uilogic.enums
-
-enum class TextFieldValidationState {
-    Empty,
-    Valid,
-    Invalid
-}
-
-interface TextFieldValidationInterface {
-    val asTextFieldState: TextFieldValidationState
-}
