--- common/src/commonMain/kotlin/me/kevinschildhorn/common/deprecated/uilogic/enums/TrailingIconState.kt
@@ -1,8 +1,0 @@
-package me.kevinschildhorn.common.deprecated.uilogic.enums
-
-enum class TrailingIconState {
-    NONE,
-    CLEAR_TEXT,
-    ERROR,
-    CHECKMARK,
-}
