+++ common/src/commonMain/kotlin/me/kevinschildhorn/common/design/atomic/atoms/Atom.kt
@@ -1,0 +1,7 @@
+package me.kevinschildhorn.common.design.atomic.atoms
+
+import me.kevinschildhorn.common.design.atomic.atoms.interfaces.AtomType
+
+abstract class Atom {
+    abstract val type: AtomType
+}
