+++ common/src/commonMain/kotlin/me/kevinschildhorn/common/design/atomic/atoms/ImageAtom.kt
@@ -1,0 +1,12 @@
+package me.kevinschildhorn.common.design.atomic.atoms
+
+import me.kevinschildhorn.common.design.atomic.atoms.interfaces.AtomType
+import me.kevinschildhorn.common.design.atomic.atoms.interfaces.ResourceAtom
+import me.kevinschildhorn.common.design.color.SharedColor
+
+class ImageAtom<T>(
+    color: SharedColor,
+    override val image: T,
+) : Atom(), ResourceAtom<T> {
+    override val type: AtomType = AtomType.IMAGE
+}
