+++ common/src/commonMain/kotlin/me/kevinschildhorn/common/design/atomic/atoms/interfaces/AtomType.kt
@@ -1,4 +1,4 @@
-package me.kevinschildhorn.common.deprecated.atomic.atoms.interfaces
+package me.kevinschildhorn.common.design.atomic.atoms.interfaces
 
 enum class AtomType {
     BUTTON,
