+++ common/src/commonMain/kotlin/me/kevinschildhorn/common/design/atomic/atoms/interfaces/ResourceAtom.kt
@@ -1,0 +1,5 @@
+package me.kevinschildhorn.common.design.atomic.atoms.interfaces
+
+interface ResourceAtom<T> {
+    val image: T
+}
